move social media components
