+++ extensions/amp-soundcloud/1.0/storybook/Basic.js
@@ -1,10 +1,10 @@
 import {boolean, color, text, withKnobs} from '@storybook/addon-knobs';
 
+import {BentoSoundcloud} from '#bento/components/bento-soundcloud/1.0/component';
+
 import * as Preact from '#preact';
 
 import {rgba2hex} from './converter';
-
-import {BentoSoundcloud} from '../component';
 
 export default {
   title: 'Soundcloud',
