+++ extensions/amp-twitter/1.0/amp-twitter.js
@@ -1,11 +1,11 @@
+import {BaseElement} from '#bento/components/bento-twitter/1.0/base-element';
+
 import {userAssert} from '#core/assert';
 import {htmlFor} from '#core/dom/static-template';
 
 import {isExperimentOn} from '#experiments';
 
 import {AmpPreactBaseElement, setSuperClass} from '#preact/amp-base-element';
-
-import {BaseElement} from './base-element';
 
 import {getBootstrapBaseUrl, getBootstrapUrl} from '../../../src/3p-frame';
