+++ src/bento/components/bento-facebook/1.0/bento-facebook.js
@@ -1,2 +1,3 @@
 import {defineElement} from './web-component';
+
 defineElement();
