+++ src/bento/components/bento-facebook/1.0/component.type.js
