+++ src/bento/components/bento-facebook/1.0/web-component.js
