+++ src/bento/components/bento-instagram/1.0/base-element.js
