+++ src/bento/components/bento-instagram/1.0/component.type.js
