+++ src/bento/components/bento-social-share/1.0/component.type.js
