+++ src/bento/components/bento-social-share/1.0/social-share-svgs.js
