+++ src/bento/components/bento-social-share/OWNERS
@@ -1,0 +1,14 @@
+// For an explanation of the OWNERS rules and syntax, see:
+// https://github.com/ampproject/amp-github-apps/blob/main/owners/OWNERS.example
+
+{
+  rules: [
+    {
+      owners: [
+        {name: 'ampproject/wg-components'},
+        {name: 'ithinkihaveacat'},
+        {name: 'ryancopley'},
+      ],
+    },
+  ],
+}
