+++ src/bento/components/bento-soundcloud/1.0/base-element.js
