+++ src/bento/components/bento-soundcloud/1.0/component.js
