+++ src/bento/components/bento-twitter/1.0/component.type.js
